Run tailored data extraction on new form
We use to use a trick to determine if a form is new or not to run the
extraction. It has a limitation: if we create a form before the data
needed for the extraction are saved at the proper place, then we
couldn't get those data back.

Now, there is a "NEW" status on the forms which gets updated only when
the form gets saved. Everytime we load a "NEW" form, we clean previously
extracted data and then we extract the most recent tailored data.